Allow extracting and modifying the orbital elements of a TLE (#9)
* Allow extracting and modifying the orbital elements of a TLE

The naming used in the original SGP4 sources can be quite baroque. This is
common for mathematically intensive code, where names often follow the symbols
used within a standard mathematical formulation, but without the corresponding
formulae to reference it can be quite difficult to know what a parameter
actually is. Fortunately we are not bound to the naming that SGP4 uses, as we
only need to be binary-compatible with the C++ implementation. This commit
renames some of the fields -- namely, those that relate to the TLE orbital
elements and a few others -- to make their meaning more obvious.

In the public API, the `TwoLineElement` structure now supports extracting and
setting these values in a type safe way. The structure uses the `uom` crate to
provide dimensional analysis, which can help avoid unit-of-measure errors.

To close the loop TLEs can be modified via setter methods, allowing simulators
to adjust the orbital elements in response to events in their scenario and
propagate the new orbit.

* Change visibility of `_sys` types to `pub(crate)`