Make the StateVector::new constructor public (#10)
StateVector is normally returned by the SGP4 propagation routine, but it
can also be used to extract the Keplerian elements when an object's
position and velocity are known at a given epoch. Exposing the
constructor allows users to skip SGP4 propagation or to integrate with
other propagators when this information is available.